MAGE-1394 Reorganize command classes
