+++ Console/Command/Replica/AbstractReplicaCommand.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Algolia\AlgoliaSearch\Console\Command;
+namespace Algolia\AlgoliaSearch\Console\Command\Replica;
+
+use Algolia\AlgoliaSearch\Console\Command\AbstractStoreCommand;
 
 abstract class AbstractReplicaCommand extends AbstractStoreCommand
 {
